Adapt button size to make the tests pass
