+++ app/src/main/res/values/strings.xml
@@ -73,7 +73,7 @@
     <string name="max_item_quantity_text">There are only %1$s %2$s available !</string>
     <string name="go_to_admin_item_list">go to admin item list</string>
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="admin_hub_btn_item_text">Item Request</string>
+    <string name="admin_hub_btn_item_text">Item</string>
     <string name="admin_hub_btn_zone_text">Zone</string>
     <string name="admin_hub_btn_event_text">Event</string>
     <string name="admin_hub_btn_user_text">User</string>
